requirements.txt
awscli==1.36.23
bcrypt==4.2.1
blinker==1.9.0
boto3==1.35.82
botocore==1.35.82
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
docutils==0.16
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
pillow==11.0.0
pyasn1==0.6.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.4
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
